feat: ignore packages incompatible with catalyst
