+++ go/internal/multipeer-connectivity-driver/bridge_unsupported.go
@@ -1,4 +1,4 @@
-// +build !darwin !cgo
+// +build !darwin darwin,!cgo catalyst
 
 package mc
