+++ go/internal/multipeer-connectivity-driver/driver/MCManager.m
